Clone the CairoSVG repo and do basic tests
Note that because `pytest-isort` is not available on conda-forge, the
test framework from the CairoSVG repo cannot be reused.  This tests the
recipe by putting all the test files that the source repo expected to
pass for this release through the `cairosvg` command line and seeing if
a non-zero-length file was created.  If any of the files failed to be
converted, the recipe fails.